chore: components reorganized into the same file
